cyberos-order-project/backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-cors-headers==3.13.0
whitenoise==5.3.0
django-environ==0.8.1